// src/app/features/transfer-channel/napas-configuration/search-dispute/modal/enum.ts
export enum typeTransactionDisputeEnum {
    REQUEST = 'REQUEST',
    RESPONSE = 'RESPONSE'
}


export enum transactionTypeEnum {
    ISS = 'ISS',
    ACQ = 'ACQ'
}

export enum approveDisputeStatusEnum {
    REJECTED = 'REJECTED',
    APPROVED = 'APPROVED',
    WAITING = 'WAITING',
    OPEN = 'OPEN',
    RESPONDED = 'RESPONDED'
}